Move code into proper folders
