+++ src/sensor/magneto/magneto1_4.c
