+++ src/sensor/magneto/magneto1_4.h
